@@ -8,8 +8,11 @@
     "lint": "vue-cli-service lint"
   },
   "dependencies": {
+    "bootstrap": "^5.1.1",
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vue": "^3.2.1",
+    "vue-resource": "^1.5.3",
+    "vue-router": "^4.0.11"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
