@@ -1,8 +1,12 @@
-import Vue from 'vue'
+// Main file for router functionality
+
+import { createApp } from 'vue'
 import App from './App.vue'
+import router from './router'
 
-Vue.config.productionTip = false
 
-new Vue({
-  render: h => h(App),
-}).$mount('#app')
+import "bootstrap/dist/js/bootstrap.bundle";
+
+import "bootstrap/dist/css/bootstrap.min.css";
+
+createApp(App).use(router).mount('#app')
